refactor(audit): move host signer keypair into mvm_hostd::audit::host_keypair with cli shim
