+++ crates/mvm-hostd/src/audit/mod.rs
@@ -2,4 +2,5 @@
 //! keypair, plan persistence, and the checkpoint bind helpers. Library API so
 //! both the CLI and fleet consumers emit identical chain entries.
 
+pub mod host_keypair;
 pub mod plan_persist;
